Regressions
Attempted a shitload of regressions but we have an endogeneity problem
with independent expenditures, where the coefficients have the opposite
signs since spending for a candidate would be induced if they were less
likely to win. The very last regression (openresife) is probably the
best one theoretically that I could come up with, where I try to control
for this using the spread of the last election. It doesn't do anything
:( I'm pretty stuck on what else to try